Added in the basic logged in menu
I added in the basic features to the logged in menu. The user can
currently log in and out. The database still needs to be made and the
validation will need to be added. The socket messages will need to be
created for when a user does anything on any menu and sends a message to
the server. Currently validated with regular expressions.