Fixed menu issues
I managed to fix the inner menu issues within my code. The code would
not allow user input after the "Display a course" menu option failed. I
added in some code that would send a valid failure response which what
was causing the issue as the user input would not be sent to the server
because it could not be flushed hence an error would occur.